fix(issue #119): Escape empty brackets [] when present in query parameters (#121)
* fix(issue #119): Escape empty brackets [] when present

If the url has query parameters and they contain empty brackets, they need to be escaped for the callback function in HttpSteps to complete normally. As done by escapeBrackets in MockFaster.
If the given step, has empty brackets but in the url path this time. Mockserver won't be able to compile the path url to a regex and crash. But for path url you can decode the empty brackets into their UTF-8 encoding, i.e. %5B for [ and %5D for ], and mockserver will still match the request with the mock.
This method of using UTF-8 encoding for the brackets didn't work  for the query parameters, otherwise Mockserver won't be able to match the mock parameter with the real one, so we have to escape them.

Co-authored-by: brian-mulier <69139631+brian-mulier@users.noreply.github.com>